move setting tabs out to seperate class and files
