+++ idarling/interface/dialogs.py
@@ -28,6 +28,7 @@ from PyQt5.QtWidgets import (QDialog, QHBoxLayout, QVBoxLayout, QGridLayout,
 from ..shared.commands import GetRepositories, GetBranches, \
     NewRepository, NewBranch
 from ..shared.models import Repository, Branch
+from .tabs import TabCfgServer, TabCfgGeneral
 
 logger = logging.getLogger('IDArling.Interface')
 
@@ -444,10 +445,10 @@ class SettingsDialog(QDialog):
         # Settings Tabs
         tabs = QTabWidget(self)
         # General Settings tab
-        tab_general = self.__tab_general(tabs)
+        tab_general = TabCfgGeneral(self, tabs)
         tabs.addTab(tab_general, "General Settings")
         # Network Settings tab
-        tab_servers = self.__tab_servers(tabs)
+        tab_servers = TabCfgServer(self, tabs)
         tabs.addTab(tab_servers, "Network Settings")
 
         layout.addWidget(tabs)
@@ -455,264 +456,6 @@ class SettingsDialog(QDialog):
         saveButton = QPushButton("Close")
         saveButton.clicked.connect(self.accept)
         layout.addWidget(saveButton)
-
-    def __tab_general(self, parent):
-        """
-        Initialize General Settings in the Settings dialog
-
-        :param: parent: Parent QTabWidget
-        :return: a QWidget containing general settings
-        """
-        # TODO: code cleaning
-        assert (type(parent) == QTabWidget)
-
-        tab = QWidget(parent)
-        # Set general layout
-        layout = QFormLayout(tab)
-        layout.setFormAlignment(Qt.AlignVCenter)
-
-        # Add widgets and restore settings
-
-        # NavbarColorizer Checkbox
-        display = "Disable users display in the navigation bar"
-        noNavbarColorizerCheckbox = QCheckBox(display)
-        layout.addRow(noNavbarColorizerCheckbox)
-
-        def noNavbarColorizerActionTriggered():
-            self._plugin.interface.painter.noNavbarColorizer = \
-                noNavbarColorizerCheckbox.isChecked()
-
-        noNavbarColorizerCheckbox.toggled.connect(
-            noNavbarColorizerActionTriggered
-        )
-        checked = self._plugin.interface.painter.noNavbarColorizer
-        noNavbarColorizerCheckbox.setChecked(checked)
-
-        # Notifications Checkbox
-        display = "Disable notifications"
-        noNotificationsCheckbox = QCheckBox(display)
-        layout.addRow(noNotificationsCheckbox)
-
-        def noNotificationsActionToggled():
-            self._plugin.interface.painter.noNotifications = \
-                noNotificationsCheckbox.isChecked()
-
-        noNotificationsCheckbox.toggled.connect(noNotificationsActionToggled)
-        checked = self._plugin.interface.painter.noNotifications
-        noNotificationsCheckbox.setChecked(checked)
-
-        # Color settings
-        color_settings = self.__tab_general_color_settings(tab)
-        layout.addWidget(color_settings)
-
-        # Debug Level settings
-        debugLevelLabel = QLabel("Log Level: ")
-        debugLevelComboBox = QComboBox()
-        debugLevelComboBox.addItem("CRITICAL", logging.CRITICAL)
-        debugLevelComboBox.addItem("ERROR", logging.ERROR)
-        debugLevelComboBox.addItem("WARNING", logging.WARNING)
-        debugLevelComboBox.addItem("INFO", logging.INFO)
-        debugLevelComboBox.addItem("DEBUG", logging.DEBUG)
-
-        def debugLevelInitialized():
-            from idarling.plugin import logger
-
-            index = debugLevelComboBox.findData(logger.getEffectiveLevel())
-            debugLevelComboBox.setCurrentIndex(index)
-
-        debugLevelInitialized()
-
-        def debugLevelActivated(index):
-            from idarling.plugin import logger
-
-            level = debugLevelComboBox.itemData(index)
-            logger.setLevel(level)
-            self._plugin.config["level"] = level
-            self._plugin.save_config()
-
-        debugLevelComboBox.activated.connect(debugLevelActivated)
-        layout.addRow(debugLevelLabel, debugLevelComboBox)
-
-        return tab
-
-    def __tab_servers(self, parent):
-        """
-        Initialize server settings in the Settings dialog
-        :param parent: QTabWidget
-        :return: a QWidget containing general settings
-        """
-        assert (type(parent) == QTabWidget)
-
-        # TODO: code cleaning
-        tab = QWidget(parent)
-        layout = QVBoxLayout(tab)
-
-        topWidget = QWidget(tab)
-        topLayout = QHBoxLayout(topWidget)
-        layout.addWidget(topWidget)
-
-        servers = self._plugin.config["servers"]
-        self._serversTable = QTableWidget(len(servers), 2, self)
-        self._serversTable.setHorizontalHeaderLabels(("Servers", ""))
-        for i, server in enumerate(servers):
-            item = QTableWidgetItem('%s:%d' % (server["host"], server["port"]))
-            item.setData(Qt.UserRole, server)
-            item.setFlags(item.flags() & ~Qt.ItemIsEditable)
-            if self._plugin.network.server == server:
-                item.setFlags((item.flags() & ~Qt.ItemIsSelectable))
-            self._serversTable.setItem(i, 0, item)
-
-            checkbox = QTableWidgetItem()
-            state = Qt.Unchecked if server["no_ssl"] else Qt.Checked
-            checkbox.setCheckState(state)
-            checkbox.setFlags((checkbox.flags() & ~Qt.ItemIsEditable))
-            checkbox.setFlags((checkbox.flags() & ~Qt.ItemIsUserCheckable))
-            if self._plugin.network.server == server:
-                checkbox.setFlags((checkbox.flags() & ~Qt.ItemIsSelectable))
-            self._serversTable.setItem(i, 1, checkbox)
-
-        horizontalHeader = self._serversTable.horizontalHeader()
-        horizontalHeader.setSectionsClickable(False)
-        horizontalHeader.setSectionResizeMode(0, QHeaderView.Stretch)
-        horizontalHeader.setSectionResizeMode(1, QHeaderView.ResizeToContents)
-        self._serversTable.verticalHeader().setVisible(False)
-        self._serversTable.setSelectionBehavior(QTableWidget.SelectRows)
-        self._serversTable.setSelectionMode(QTableWidget.SingleSelection)
-        self._serversTable.itemClicked.connect(self._server_clicked)
-        self._serversTable.itemDoubleClicked.connect(
-            self._server_double_clicked)
-        minSZ = self._serversTable.minimumSize()
-        self._serversTable.setMinimumSize(300, minSZ.height())
-        topLayout.addWidget(self._serversTable)
-
-        buttonsWidget = QWidget(self)
-        buttonsLayout = QVBoxLayout(buttonsWidget)
-
-        # Add server button
-        self._addButton = QPushButton("Add Server")
-        self._addButton.clicked.connect(self._add_button_clicked)
-        buttonsLayout.addWidget(self._addButton)
-
-        # Edit server button
-        self._editButton = QPushButton("Edit Server")
-        self._editButton.setEnabled(False)
-        self._editButton.clicked.connect(self._edit_button_clicked)
-        buttonsLayout.addWidget(self._editButton)
-
-        # Delete server button
-        self._deleteButton = QPushButton("Delete Server")
-        self._deleteButton.setEnabled(False)
-        self._deleteButton.clicked.connect(self._delete_button_clicked)
-        buttonsLayout.addWidget(self._deleteButton)
-
-        # Cancel button
-        topLayout.addWidget(buttonsWidget)
-
-        bottomWidget = QWidget(tab)
-        bottomLayout = QHBoxLayout(bottomWidget)
-        layout.addWidget(bottomWidget)
-
-        def update_keep_alive():
-            self._plugin.save_config()
-            cnt = self._plugin.config["keep"]["cnt"]
-            intvl = self._plugin.config["keep"]["intvl"]
-            idle = self._plugin.config["keep"]["idle"]
-            if self._plugin.network.client:
-                self._plugin.network.client.set_keep_alive(cnt, intvl, idle)
-
-        keepWidget = QWidget(tab)
-        keepLayout = QFormLayout(keepWidget)
-        keepCntLabel = QLabel("Keep-Alive Count: ")
-        keepCntSpinBox = QSpinBox(keepWidget)
-        keepCntSpinBox.setRange(0, 86400)
-        keepCntSpinBox.setValue(self._plugin.config["keep"]["cnt"])
-        keepCntSpinBox.setSuffix(" packets")
-
-        def keepCntSpinBoxChanged(cnt):
-            self._plugin.config["keep"]["cnt"] = cnt
-            update_keep_alive()
-
-        keepCntSpinBox.valueChanged.connect(keepCntSpinBoxChanged)
-        keepLayout.addRow(keepCntLabel, keepCntSpinBox)
-
-        keepIntvlLabel = QLabel("Keep-Alive Interval: ")
-        keepIntvlSpinBox = QSpinBox(keepWidget)
-        keepIntvlSpinBox.setRange(0, 86400)
-        keepIntvlSpinBox.setValue(self._plugin.config["keep"]["intvl"])
-        keepIntvlSpinBox.setSuffix(" seconds")
-
-        def keepIntvlSpinBoxChanged(intvl):
-            self._plugin.config["keep"]["intvl"] = intvl
-            update_keep_alive()
-
-        keepIntvlSpinBox.valueChanged.connect(keepIntvlSpinBoxChanged)
-        keepLayout.addRow(keepIntvlLabel, keepIntvlSpinBox)
-
-        keepIdleLabel = QLabel("Keep-Alive Idle: ")
-        keepIdleSpinBox = QSpinBox(keepWidget)
-        keepIdleSpinBox.setRange(0, 86400)
-        keepIdleSpinBox.setValue(self._plugin.config["keep"]["idle"])
-        keepIdleSpinBox.setSuffix(" seconds")
-
-        def keepIdleSpinBoxChanged(idle):
-            self._plugin.config["keep"]["idle"] = idle
-            update_keep_alive()
-
-        keepIdleSpinBox.valueChanged.connect(keepIdleSpinBoxChanged)
-        keepLayout.addRow(keepIdleLabel, keepIdleSpinBox)
-        bottomLayout.addWidget(keepWidget)
-
-        return tab
-
-    def __tab_general_color_settings(self, parent):
-        """
-        Initialize color settings in the General Settings tab
-        :param parent: Parent QWidget
-        :return:
-        """
-        # Add User Color settings
-        colorWidget = QWidget(parent)
-        colorLayout = QHBoxLayout(colorWidget)
-
-        colorButton = QPushButton("")
-        colorButton.setFixedSize(50, 30)
-
-        def setColor(color):
-            """
-            Sets the color (if valid) as user's color
-
-            :param color: the color
-            """
-            if color.isValid():
-                r, g, b, _ = color.getRgb()
-                rgbColor = r << 16 | g << 8 | b
-                # set the color as user's color
-                self._plugin.interface.painter.color = rgbColor
-                # set the background button color
-                palette = colorButton.palette()
-                role = colorButton.backgroundRole()
-                palette.setColor(role, color)
-                colorButton.setPalette(palette)
-                colorButton.setAutoFillBackground(True)
-
-        userColor = self._plugin.interface.painter.color
-        color = QColor(userColor)
-        setColor(color)
-
-        # Add a handler on clicking color button
-        def colorButtonClicked(_):
-            color = QColorDialog.getColor()
-            setColor(color)
-
-        colorButton.clicked.connect(colorButtonClicked)
-        colorLayout.addWidget(colorButton)
-        # User name
-        self.colorLabel = QLineEdit()
-        self.colorLabel.setPlaceholderText("Name")
-        name = self._plugin.interface.painter.name
-        self.colorLabel.setText(name)
-        colorLayout.addWidget(self.colorLabel)
-        return colorWidget
 
     def _server_clicked(self, _):
         """
